+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Allow to create steady-state hypothesis from init command [#28][28]
 - Allow to set rollbacks from init command [#30][30]
 - Pass command executed to checker for compatbility [#36][36]
+- Better logging of failed discovery [chaostoolkit-lib#29][29lib]
 
 [27]: https://github.com/chaostoolkit/chaostoolkit/issues/27
 [28]: https://github.com/chaostoolkit/chaostoolkit/issues/28
@@ -29,6 +30,7 @@
 [30]: https://github.com/chaostoolkit/chaostoolkit/issues/30
 [31]: https://github.com/chaostoolkit/chaostoolkit/issues/31
 [36]: https://github.com/chaostoolkit/chaostoolkit/issues/36
+[29lib]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/29
 
 ## [0.9.0][] - 2018-01-17