+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit/chaostoolkit/compare/0.10.0...HEAD
+[Unreleased]: https://github.com/chaostoolkit/chaostoolkit/compare/0.11.0...HEAD
+
+## [0.11.0][] - 2018-02-08
+
+[0.11.0]: https://github.com/chaostoolkit/chaostoolkit/compare/0.10.0...0.11.0
 
 ### Changed